+++ app/models/reservation.rb
@@ -37,6 +37,43 @@ class Reservation < ApplicationRecord
   scope :instalment, -> { where(state: INSTALMENT) }
   scope :paid, -> { where(state: PAID) }
 
+  # These are rights that may become visible over time, with the possibility of distinguishing between a right that's
+  # currently able to be used or one that's coming soon. These also match i18n values in config/locales
+  def active_rights
+    [].tap do |rights|
+      # Hold these memberships in memory to avoid hitting the database a lot
+      memberships_held = Membership.where(id: orders.select(:membership_id))
+
+      rights << "rights.attend" if memberships_held.any?(&:can_attend?)
+      rights << "rights.site_selection" if memberships_held.any?(&:can_site_select?)
+
+      now = DateTime.now
+      if now < $nomination_opens_at
+        if memberships_held.any?(&:can_nominate?)
+          rights << "rights.hugo.nominate_soon"
+          rights << "rights.retro_hugo.nominate_soon"
+        end
+      elsif now.between?($nomination_opens_at, $voting_opens_at)
+        if memberships_held.any?(&:can_nominate?) && memberships_held.none?(&:can_vote?)
+          rights << "rights.hugo.nominate_only"
+          rights << "rights.retro_hugo.nominate_only"
+        elsif memberships_held.any?(&:can_nominate?)
+          rights << "rights.hugo.nominate"
+          rights << "rights.retro_hugo.nominate"
+        end
+      elsif now.between?($voting_opens_at, $hugo_closed_at)
+        if memberships_held.any?(&:can_vote?)
+          rights << "rights.hugo.vote"
+          rights << "rights.retro_hugo.vote"
+        end
+      end
+    end
+  end
+
+  def can_nominate?
+    Membership.can_nominate.where(id: orders.select(:membership_id)).exists?
+  end
+
   def paid?
     state == PAID
   end